refactor(cli): No private modules
